+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [Unreleased]
 
+### Added
+- Speculative Sampling
+
 
 ## [0.1.0] - 2024-05-20